Fix clients not being able to shoot
This is actually doing a workaround for an issue we've had for a while
but hasn't mattered until now. We're not actually syncing the player
state, previously "is_active", and we're using that state to check if
the player should be able to fire. Previously, is_active defaulted to
true, meaning that on the server, it would think all clients are always
active. But the new default state is FROZEN, and since it's not syncing
across the proper new state when it becomes NORMAL, it didn't let them
shoot because it thought they were still frozen. Eventually we should
probably sync that over, and in a way that can't be cheated around